docs: decaying trajectories
